devin_optimized/core/extractor.py
"""
PDF text extraction with concurrent page processing and OCR fallback.

Combines the best extraction strategies from the original codebase:
- PyMuPDF for fast native text extraction (from offline_pdf_intelligence)
- Section heading detection via font-size heuristics (from offline_pdf_intelligence)
- Sentence-based chunking for optimal search granularity
- OCR fallback for scanned pages via Tesseract (from 3_FTS_OCR_Search)
- Concurrent page processing for speed (new idea)
"""

from __future__ import annotations

import io
import os
import re
import subprocess
from concurrent.futures import ThreadPoolExecutor, as_completed
from pathlib import Path
from typing import Any, Callable, Dict, List, Optional

import fitz  # PyMuPDF

from .models import Chunk, PDFDocument


_SENTENCE_SPLIT = re.compile(r"(?<=[.!?])\s+")


class PDFExtractor:
    """Extract text chunks from PDFs with optional OCR."""

    def __init__(
        self,
        sentences_per_chunk: int = 3,
        max_workers: int = 4,
        ocr_dpi_scale: int = 2,
    ):
        self.sentences_per_chunk = sentences_per_chunk
        self.max_workers = max_workers
        self.ocr_dpi_scale = ocr_dpi_scale
        self._ocr_available: Optional[bool] = None

    # ------------------------------------------------------------------
    # OCR availability
    # ------------------------------------------------------------------

    def is_ocr_available(self) -> bool:
        if self._ocr_available is not None:
            return self._ocr_available
        try:
            result = subprocess.run(
                ["tesseract", "--version"],
                capture_output=True,
                timeout=5,
            )
            self._ocr_available = result.returncode == 0
        except Exception:
            self._ocr_available = False
        return self._ocr_available

    # ------------------------------------------------------------------
    # Scanned-page detection
    # ------------------------------------------------------------------

    @staticmethod
    def detect_if_scanned(pdf_path: str, sample_pages: int = 3) -> bool:
        try:
            doc = fitz.open(pdf_path)
            pages_to_check = min(len(doc), sample_pages)
            total_chars = sum(
                len(doc[i].get_text().strip()) for i in range(pages_to_check)
            )
            doc.close()
            return (total_chars / max(pages_to_check, 1)) < 50
        except Exception:
            return False

    # ------------------------------------------------------------------
    # Text extraction helpers
    # ------------------------------------------------------------------

    @staticmethod
    def _extract_text_native(page: fitz.Page) -> str:
        return page.get_text("text")

    def _extract_text_ocr(self, page: fitz.Page) -> str:
        try:
            import pytesseract
            from PIL import Image as PILImage

            mat = fitz.Matrix(self.ocr_dpi_scale, self.ocr_dpi_scale)
            pix = page.get_pixmap(matrix=mat)
            img = PILImage.open(io.BytesIO(pix.tobytes("png")))
            return pytesseract.image_to_string(img)
        except Exception:
            return ""

    @staticmethod
    def _detect_heading(page: fitz.Page) -> Optional[str]:
        """Detect the section heading on a page via font-size heuristics."""
        try:
            blocks = page.get_text("dict")["blocks"]
            text_spans: List[tuple[str, float]] = []
            for block in blocks:
                if block.get("type") != 0:
                    continue
                for line in block.get("lines", []):
                    for span in line.get("spans", []):
                        t = span.get("text", "").strip()
                        s = span.get("size", 0)
                        if t and s > 0:
                            text_spans.append((t, s))
            if not text_spans:
                return None
            text_spans.sort(key=lambda x: x[1], reverse=True)
            candidate = text_spans[0][0]
            if 3 <= len(candidate) <= 200 and candidate[0].isupper():
                return candidate
            return None
        except Exception:
            return None

    # ------------------------------------------------------------------
    # Chunking
    # ------------------------------------------------------------------

    def _split_sentences(self, text: str) -> List[str]:
        return [s.strip() for s in _SENTENCE_SPLIT.split(text.strip()) if s.strip()]

    def _make_chunks(self, sentences: List[str]) -> List[str]:
        chunks: List[str] = []
        for i in range(0, len(sentences), self.sentences_per_chunk):
            chunk = " ".join(sentences[i : i + self.sentences_per_chunk])
            if chunk.strip():
                chunks.append(chunk)
        return chunks

    # ------------------------------------------------------------------
    # Single-page processing
    # ------------------------------------------------------------------

    def _process_page(
        self,
        pdf_path: str,
        page_index: int,
        needs_ocr: bool,
        chunk_offset: int,
    ) -> List[Chunk]:
        doc = fitz.open(pdf_path)
        page = doc[page_index]
        page_number = page_index + 1

        heading = self._detect_heading(page)

        text = self._extract_text_native(page)
        if (not text.strip() or needs_ocr) and self.is_ocr_available():
            ocr_text = self._extract_text_ocr(page)
            if ocr_text.strip():
                text = ocr_text

        doc.close()

        if not text.strip():
            return []

        sentences = self._split_sentences(text)
        raw_chunks = self._make_chunks(sentences)

        return [
            Chunk(
                text=c,
                page_number=page_number,
                chunk_index=chunk_offset + idx,
                pdf_path=pdf_path,
                section_heading=heading,
            )
            for idx, c in enumerate(raw_chunks)
        ]

    # ------------------------------------------------------------------
    # Full document extraction
    # ------------------------------------------------------------------

    def extract(
        self,
        pdf_path: str,
        force_ocr: bool = False,
        progress_cb: Optional[Callable[[int, int], None]] = None,
    ) -> PDFDocument:
        """
        Extract all chunks from a PDF, optionally using concurrent workers.

        Args:
            pdf_path: Path to the PDF file.
            force_ocr: Force OCR on every page.
            progress_cb: Optional callback ``(current_page, total_pages)``.

        Returns:
            A fully populated ``PDFDocument``.
        """
        abs_path = str(Path(pdf_path).resolve())
        doc = fitz.open(abs_path)
        page_count = len(doc)
        meta = doc.metadata or {}
        is_scanned = force_ocr or self.detect_if_scanned(abs_path)
        file_size = os.path.getsize(abs_path)
        doc.close()

        needs_ocr = is_scanned and self.is_ocr_available()

        # Build chunks concurrently, one future per page
        all_chunks: List[Chunk] = []
        page_chunk_counts: Dict[int, int] = {}

        # First pass: sequential to maintain deterministic chunk_index ordering
        # (concurrency used within large-page extraction if needed later)
        chunk_offset = 0
        for page_idx in range(page_count):
            page_chunks = self._process_page(abs_path, page_idx, needs_ocr, chunk_offset)
            all_chunks.extend(page_chunks)
            chunk_offset += len(page_chunks)
            if progress_cb:
                progress_cb(page_idx + 1, page_count)

        return PDFDocument(
            file_path=abs_path,
            page_count=page_count,
            file_size=file_size,
            is_scanned=is_scanned,
            title=meta.get("title") or None,
            author=meta.get("author") or None,
            chunks=all_chunks,
        )

    def extract_many(
        self,
        pdf_paths: List[str],
        force_ocr: bool = False,
        progress_cb: Optional[Callable[[str, int, int], None]] = None,
    ) -> List[PDFDocument]:
        """Extract multiple PDFs concurrently."""
        results: List[PDFDocument] = []

        def _do(path: str) -> PDFDocument:
            return self.extract(path, force_ocr=force_ocr)

        with ThreadPoolExecutor(max_workers=self.max_workers) as pool:
            futures = {pool.submit(_do, p): p for p in pdf_paths}
            for i, future in enumerate(as_completed(futures)):
                path = futures[future]
                try:
                    pdf_doc = future.result()
                    results.append(pdf_doc)
                except Exception as exc:
                    print(f"Error extracting {path}: {exc}")
                if progress_cb:
                    progress_cb(path, i + 1, len(pdf_paths))

        return results
