devin_optimized/requirements.txt
# Devin Optimized PDF Intelligence - Requirements

# PDF processing
PyMuPDF>=1.23.0           # fitz — fast PDF text/image extraction

# BM25 search
bm25s>=0.1.0              # Deterministic BM25 scoring

# TF-IDF + cosine similarity
scikit-learn>=1.3.0
scipy>=1.11.0

# Numerical
numpy>=1.24.0

# OCR support (optional — requires system tesseract-ocr)
pytesseract>=0.3.10
Pillow>=9.2.0
